Refactored forms to use Flask-Bootstrap
Changed cancel buttons on forms to be links because of
Flask-Bootstrap behavior requiring the required form fields
to be filled out before cancelling. The cancel button did
have a type of submit to be able to add code for the button
to the view. Because the button is now a link, it no longer
displays a flash message that an action has been cancelled,
which is OK because the user will understand it is
cancelled when it goes to the previous page.

Updated titles on pages.