Switch to PostgreSQL database
Switched from SQLite Database to PostgreSQL database since it is
better for web apps.

Renamed user table to users because PostgreSQL already had a
built-in user table.

Updated some field lengths because test data had more characters
in fields than allowed, and it did not give any errors with the
old SQLite database.

Added the Vagrantfile and pg_config.sh which creates the
PostgreSQL database.

Put a constraint on collection path to be unique.